Convert Paragraph to TypeScript
[#178817898]
